% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{calculate.purity}
\alias{calculate.purity}
\title{Calculate Gene Purity (Internal)}
\usage{
calculate.purity(seu, gene.pct, neighbor, ncores = 10)
}
\arguments{
\item{seu}{Seurat object}

\item{gene.pct}{Data frame with gene filters}

\item{neighbor}{K for KNN}

\item{ncores}{Cores}
}
\description{
Calculate Gene Purity (Internal)
}
