% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_functions.R
\name{calculate.silhouette}
\alias{calculate.silhouette}
\title{Calculate Silhouette Scores (Internal)}
\usage{
calculate.silhouette(seu, purity, ncores = 10)
}
\arguments{
\item{seu}{Seurat object}

\item{purity}{Output from calculate.purity}

\item{ncores}{Cores}
}
\description{
Calculate Silhouette Scores (Internal)
}
